LayoutTests/http/tests/storageAccess/deny-storage-access-under-opener-if-auto-dismiss-expected.txt

Tests that a cross-origin window from a prevalent domain with non-recent user interaction doesn't get storage access under its opener if it just loads and auto-dismisses.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Cookie created.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
Should not receive first-party cookie.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named ''.
Client-side document.cookie:
